backend/requirements.txt
requests==2.26.0
django==3.0.0
djangorestframework==3.12.4
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
djangorestframework-simplejwt==5.0.0
django-filter==21.1